+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
+## Added
+- Added `PebbleApiException::getResponse()` to retrieve the original raw API response when a pin push attempt fails.
 
 ## [1.1.0] - 2015-09-10
 ### Added